takepatch.c	-S - save directories even if errors (for debugging). slib.c  	Don't error on empty diffs. rechksum.c	Add -v to list status as checking.
bk: 37c8b56ekkl9LvKPHwPTZl7K41PsZA

+++ src/takepatch.c
@@ -30,6 +30,7 @@ usage: takepatch [-cFiv] [-f file]\n\n\
     -F		(fast) do rebuild id cache when creating files\n\
     -f<file>	take the patch from <file> and do not save it\n\
     -i		initial patch, create a new repository\n\
+    -S		save RESYNC and or PENDING directories even if errors\n\
     -v		verbose level, more is more verbose, -vv is suggested.\n\n";
 
 #define	CLEAN_RESYNC	1	/* blow away the RESYNC dir */
@@ -62,6 +63,7 @@ int	fileNum;	/* counter for the Nth init/diff file */
 patch	*patchList = 0;	/* list of patches for a file, list len == fileNum */
 int	conflicts;	/* number of conflicts over all files */
 int	newProject;	/* command line option to create a new repository */
+int	saveDirs;	/* save directories even if errors */
 MDBM	*idDB;		/* key to pathname database, set by init or rebuilt */
 MDBM	*goneDB;	/* key to gone database */
 delta	*gca;		/* The oldest parent found in the patch */
@@ -87,12 +89,12 @@ main(int ac, char **av)
 
 	input = stdin;
 	debug_main(av);
-	while ((c = getopt(ac, av, "acFf:iqsv")) != -1) {
+	while ((c = getopt(ac, av, "acFf:iqsSv")) != -1) {
 		switch (c) {
 		    case 'q':
 		    case 's':
-		    /* Ignored for option consistency.  */
-		    break;
+			/* Ignored for option consistency.  */
+			break;
 		    case 'a': resolve++; break;
 		    case 'c': noConflicts++; break;
 		    case 'F': fast++; break;
@@ -105,6 +107,7 @@ main(int ac, char **av)
 			    }
 			    break;
 		    case 'i': newProject++; break;
+		    case 'S': saveDirs++; break;
 		    case 'v': echo++; flags &= ~SILENT; break;
 		    default: goto usage;
 		}
@@ -702,6 +705,7 @@ apply:
 				    	perror("get");
 					cleanup(CLEAN_RESYNC);
 				}
+				sccs_restart(s);
 				if (echo > 6) {
 					char	buf[MAXPATH];
 
@@ -1258,6 +1262,11 @@ rebuild_id(char *id)
 void
 cleanup(int what)
 {
+	if (saveDirs) {
+		fprintf(stderr, "takepatch: neither directory removed.\n");
+		SHOUT2();
+		exit(1);
+	}
 	if (what & CLEAN_RESYNC) {
 		assert(exists("RESYNC"));
 		system("/bin/rm -rf RESYNC");
